requirements.txt
pinecone
openai
pypdf
python-dotenv
